requirements.txt
# Core AI + Speech
openai
faster-whisper
SpeechRecognition
pyaudio
langdetect
gTTS
playsound
pvporcupine 
sounddevice
onnxruntime
openwakeword

# Dashboard
Flask

# Telegram Bot
python-telegram-bot==20.6
telegram

# System, Files, Scheduling
psutil
watchdog
pyyaml
python-dotenv
pytz
dateparser

# MQTT (optional – comment if unused)
paho-mqtt

# Zigbee / Home Assistant Support (optional – comment if unused)
zigpy
bellows
zigpy-znp
zha-quirks
RPi.GPIO; platform_system == "Linux"
gpiozero; platform_system == "Linux"

# FFmpeg and media support
ffmpeg-python

# File generation support (optional – future features)
python-docx
python-pptx
openpyxl
reportlab  # for basic PDFs
dotenv
PyYAML
numpy

#web-app
flask
flask-cors
